Extract Launchplane self-deploy workflow
